Re-named baselines to be better for importing. Added details to baseline logger.
